+++ includes/minishell.h
@@ -6,7 +6,7 @@
 /*   By: jvacaris <jvacaris@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/07 00:43:55 by jvacaris          #+#    #+#             */
-/*   Updated: 2021/11/21 18:50:44 by jvacaris         ###   ########.fr       */
+/*   Updated: 2021/11/21 19:16:39 by jvacaris         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -51,12 +51,12 @@ void	ft_echo(char **list);
 int		qm_error_detector(char *str);
 void	init_ms_env(char **env_vector, t_str **env_list);
 char	**env_list_to_vector(t_str **env_list);
-void	ft_lst_str_free(t_str *list);
+void	ft_lst_str_free(t_str **list);
 void	ft_lst_str_print(t_str *list);
 void	ft_lst_str_add_back(t_str **list, char *str);
 void	ft_lst_str_add_front(t_str **list, char *str);
 void	ft_lst_str_add_sorted(t_str **list, char *str);
-void	ft_lst_str_delete(t_str **list, char *str);
+void	ft_lst_str_delete(t_str **list, char *str, size_t len);
 t_str	*ft_lst_str_get_str(t_str **list, const char *str);
 void	exec_command(char **list, char **envp);
 char	*ft_strslashjoin(char const *s1, char const *s2);
@@ -83,5 +83,6 @@ void	lst_red_add_front(t_red **list, t_red *new);
 void	new_exec_command(t_red *red_node, char **env);
 void	lst_red_print(t_red *list);
 
-
+int		has_last_redirection_open(const char *str);
+int		has_pipe_redir_open(char **array);
 #endif